+++ .github/workflows/cj-action-release.yml
@@ -44,6 +44,11 @@ jobs:
             exit 1
           fi
 
+      - name: Run unit tests
+        run: |
+          python -m pip install pytest requests
+          python -m pytest runner_test.py
+
       - name: Create and Push Tag
         id: create_tag
         run: |
